chore: Move Accordion and SocialLinks components to `sections/common/` folder
