+++ src/components/layout/footer/Footer.tsx
@@ -3,7 +3,7 @@
 import ProfileImage from "@/components/sections/common/ProfileImage";
 import { profileInfo } from "@/lib/site";
 import { FiHeart } from "react-icons/fi";
-import SocialLinks from "../../ui/SocialLinks";
+import SocialLinks from "../../sections/common/SocialLinks";
 
 export default function Footer() {
   return (
